Add files via upload
So here are all the pictures for now, i put in a reference picture for the character selet, left the title of the video it is from, if you need to search it up. Do note to yourself babyboy that we want it to either hop onto the random select character (the "?" guy) or be empty, and let the user choose it for themselves. Also, if ya cant crop out the icons from the vid, just try croping it out from the character select png's. :purple_heart: